intel_gpu_top: account for per context statistics
The pipeline statistics counters represent per context values and so we
can't assume that a snapshot taken once per second will correspond to the
same context as the last snapshot.

We now read the statistics counters for every sample and before and
after each sample we read the current context id to give us a way
of detecting if the context changes.

There is now a slightly clearer separation between collecting sample
data into a fixed sized array and then handling analytics once per
second before printing/outputing the results.

Signed-off-by: Robert Bragg <robert.bragg@intel.com>