tests: perf: ensure all reports are read until timestamp
We might encounter situations where we read all the available data,
but reports about what's happening have yet to be generated. Because
we ask for periodic reports, we're better off continuously reading
reports until we see that we've read beyond the timestamp that delimits
the workload we want to measure.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>